requirements.txt
mcp
fastmcp
youtube-transcript-api
python-dotenv
requests
streamlit
openai
